Move graph processor into session_processor_default
